chore: Restructure Team form
