+++ app/(general)/teams/[id]/edit/page.tsx
@@ -1,9 +1,18 @@
-const TeamEdit = () => {
+import { redirect } from "next/navigation"
+
+import TeamForm from "@/app/(general)/teams/_components/TeamForm"
+import { auth } from "@/auth"
+import ROUTES from "@/constants/routes"
+
+const TeamEditPage = async () => {
+  const session = await auth()
+  if (!session) redirect(ROUTES.LOGIN.url)
+
   return (
-    <>
-      팀 수정 및 삭제
-    </>
+    <div>
+      <TeamForm initialData={{}} />
+    </div>
   )
 }
 
-export default TeamEdit
+export default TeamEditPage
