+++ app/(general)/teams/_components/TeamForm/FirstPage/index.tsx
@@ -3,10 +3,10 @@ import { useForm } from "react-hook-form"
 import { AiFillExclamationCircle } from "react-icons/ai"
 import { z } from "zod"
 
-import Description from "@/app/(general)/teams/create/_components/TeamCreateForm/Description"
-import basicInfoSchema from "@/app/(general)/teams/create/_components/TeamCreateForm/FirstPage/schema"
-import YoutubeDialog from "@/app/(general)/teams/create/_components/TeamCreateForm/FirstPage/YoutubeDialog"
-import Paginator from "@/app/(general)/teams/create/_components/TeamCreateForm/Paginator"
+import Description from "@/app/(general)/teams/_components/TeamForm/Description"
+import basicInfoSchema from "@/app/(general)/teams/_components/TeamForm/FirstPage/schema"
+import YoutubeDialog from "@/app/(general)/teams/_components/TeamForm/FirstPage/YoutubeDialog"
+import Paginator from "@/app/(general)/teams/_components/TeamForm/Paginator"
 import SimpleLabel from "@/components/Form/SimpleLabel"
 import { Checkbox } from "@/components/ui/checkbox"
 import { Form } from "@/components/ui/form"
