+++ app/(general)/teams/_components/TeamForm/SecondPage/MemberSessionRequiredCheckbox.tsx
@@ -2,7 +2,7 @@ import { useState } from "react"
 import { UseFormReturn } from "react-hook-form"
 import z from "zod"
 
-import { memberSessionRequiredBaseSchema } from "@/app/(general)/teams/create/_components/TeamCreateForm/SecondPage/schema"
+import { memberSessionRequiredBaseSchema } from "@/app/(general)/teams/_components/TeamForm/SecondPage/schema"
 import { Checkbox } from "@/components/ui/checkbox"
 import { Label } from "@/components/ui/label"
