+++ app/(general)/teams/_components/TeamForm/ThirdPage/index.tsx
@@ -6,10 +6,10 @@ import { FieldErrors, useForm } from "react-hook-form"
 import { AiFillExclamationCircle } from "react-icons/ai"
 import { z } from "zod"
 
-import Description from "@/app/(general)/teams/create/_components/TeamCreateForm/Description"
-import Paginator from "@/app/(general)/teams/create/_components/TeamCreateForm/Paginator"
-import { memberSessionRequiredBaseSchema } from "@/app/(general)/teams/create/_components/TeamCreateForm/SecondPage/schema"
-import UserSelect from "@/app/(general)/teams/create/_components/TeamCreateForm/ThirdPage/UserSelect"
+import Description from "@/app/(general)/teams/_components/TeamForm/Description"
+import Paginator from "@/app/(general)/teams/_components/TeamForm/Paginator"
+import { memberSessionRequiredBaseSchema } from "@/app/(general)/teams/_components/TeamForm/SecondPage/schema"
+import UserSelect from "@/app/(general)/teams/_components/TeamForm/ThirdPage/UserSelect"
 import { Form } from "@/components/ui/form"
 import API_ENDPOINTS from "@/constants/apiEndpoints"
 import fetchData from "@/lib/fetch"
